+++ src/index.js
@@ -11,17 +11,8 @@ const {
 
 const CozyUtils = require('./CozyUtils')
 const GoogleUtils = require('./GoogleUtils')
+const getAccountId = require('./helpers/getAccountId')
 const synchronizeContacts = require('./synchronizeContacts')
-
-function getAccountId() {
-  try {
-    return process.env.NODE_ENV === 'development'
-      ? 'fakeAccountId'
-      : JSON.parse(process.env.COZY_FIELDS).account
-  } catch (err) {
-    throw new Error(`You must provide 'account' in COZY_FIELDS: ${err.message}`)
-  }
-}
 
 module.exports = new BaseKonnector(start)
